chore: refactor deploy scripts
